Replace SideBar component by Bar component
FIx minor issues
