+++ app/components/SideBar/Logo/Logo.jsx
