+++ app/components/SideBar/SideBar.module.css
@@ -45,105 +45,12 @@
 	flex: 1;
 }
 
-.sidebarMenuBottomTotalBalanceShort {
-	position: relative;
-	height: 56px;
-	padding: 0;
-	cursor: pointer;
-}
-
-.sidebarMenuBottomTotalBalanceShortSeparator {
-	height: 7px;
-	margin-bottom: 15px;
-	background-image: var(--menu-arrow-up);
-	background-position: 50% 8px;
-	background-repeat: no-repeat;
-	padding: 7px 16px 0 16px;
-	background-size: 12px 6px;
-	position: relative;
-}
-
-.sidebarMenuBottomTotalBalanceShortSeparator::before {
-	content: "";
-	border-bottom: 1px solid var(--short-separator-border);
-	position: absolute;
-	bottom: 0;
-	left: 19px;
-	right: 16px;
-}
-
-.sidebarMenuBottomTotalBalanceShortSeparator.showAccounts {
-	background-color: var(--background-container-opaque);
-}
-
-.sidebarMenuBottomTotalBalanceShortName {
-	float: left;
-	color: var(--disabled-color);
-	font-size: 13px;
-	text-align: left;
-	text-decoration: none;
-	padding-left: 36px;
-	margin-left: 4px;
-	background-image: var(--sidebar-balance-icon);
-	background-repeat: no-repeat;
-	background-size: 16px 16px;
-	background-position: 16px 50%;
-}
-
-.sidebarMenuBottomTotalBalanceShortValue {
-	float: right;
-	font-family: var(--font-family-regular);
-	color: var(--sidebar-bottom-light-text-color);
-	font-size: 13px;
-	text-align: right;
-	font-weight: 600;
-	padding-right: 16px;
-}
-
-.sidebarMenuBottomTotalBalanceShortValue .balanceSmall {
-	font-size: 13px;
-}
-
-.sidebarMenuBottomLatestBlock {
-	position: relative;
-	padding: 16px 16px 16px 20px;
-	background-color: var(--info-modal-button-text);
-	font-size: 13px;
-}
-
 .sidebar.sidebarReduced .sidebarMenuBottomLatestBlock {
 	padding: 14px 0 10px 0;
 }
 
-.sidebar.sidebarReduced .sidebarMenuBottomLatestBlock .rescanButtonArea {
-	text-align: center;
-	padding-right: 0;
-	float: none;
-}
-
 .sidebar.sidebarReduced.sidebarOnBottom .sidebarMenuBottomLatestBlock {
 	display: none;
-}
-
-.sidebarMenuBottomLatestBlock-name {
-	color: var(--disabled-color);
-	font-size: 13px;
-	line-height: 10px;
-	text-align: left;
-	text-decoration: none;
-}
-
-.sidebarMenuBottomLatestBlock-number {
-	font-family: var(--font-family-monospace);
-	color: var(--sidebar-bottom-light-text-color);
-}
-
-.sidebarMenuBottomLatestBlockTime {
-	float: right;
-	color: var(--disabled-color);
-	font-size: 13px;
-	text-align: right;
-	text-decoration: none;
 }
 
 .sidebarWatchOnlyIcon {
@@ -189,14 +96,6 @@
 		margin-left: 0;
 		margin-right: auto;
 	}
-
-	.sidebarMenuBottomLatestBlock {
-		background-color: var(--sidebar-color);
-	}
-
-	.sidebarMenuBottomLatestBlock .tooltipContainer span {
-		visibility: hidden;
-	}
 }
 
 @keyframes spin {
@@ -207,8 +106,4 @@
 	to {
 		transform: rotate(360deg);
 	}
-}
-
-.rescan-button.spin {
-	animation: spin 1s linear infinite;
 }
