+++ app/constants/Decrediton.js
@@ -81,3 +81,6 @@ export const TRANSACTION_DIR_TRANSFERRED = "transfer";
 // Default name to privacy accounts needed for creation
 export const MIXED_ACCOUNT = "mixed";
 export const CHANGE_ACCOUNT = "unmixed";
+
+// MENU_LINKS_PER_ROW is the default number of menu items showed in sidebar when it's located on bottom.
+export const MENU_LINKS_PER_ROW = 4;
